requirements.txt
click==8.1.8
enum-tools==0.13.0
Flask==2.2.3
idascript==0.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
networkx==3.4.2
progressbar2==4.5.0
protobuf==6.30.2
psutil==7.0.0
Pygments==2.19.1
python-bindiff==0.3.1
python-binexport==0.3.6
python-dotenv==1.0.0
python-magic==0.4.27
python-utils==3.9.1
typing_extensions==4.13.2
Werkzeug==2.2.3
